LandMine objective functions

landmine_fitSN() is used for the module.

landmine_optim_fitSN(
  sna,
  ros,
  centreCell,
  fireSizes = 10^(2:5),
  desiredPerimeterArea = 0.004
)

landmine_optim_fitSN2(
  par,
  ros,
  centreCell,
  fireSizes = 10^(2:5),
  desiredPerimeterArea = 0.003,
  spreadProb = 0.9
)

Arguments

sna

(i.e., spawnNewActive) A numeric vector of length 4. These are the probabilities of creating spreading to 2 neighbours instead of the 1 default neighbour, each time step. The 4 values are for 4 different fire size conditions. See details in landmine_optim_burnFun().

ros

Character, specifying the file path to raster of LandMine Rate Of Spread values.

centreCell

Integer id of the centre (start) cell of ros raster. See startCells in landmine_burn1().

fireSizes

A numeric vector indicating the final size of each of the fires. See landmine_burn1().

desiredPerimeterArea

Numeric target perimeter-area ratio.

par

parameter vector of length 5

spreadProb

spread probability

Value

Summary data.table of fit results.

Details

landmine_fitSN2() is an alternative version tries the optimization using fewer parameters, to test whether a simpler version gets better/different results. Although this version was not used for the final module, we preserve it here for posterity.

See also