Wrapper function to setup cluster, export objects and load packages

Wrapper function to setup cluster, export objects and load packages

landmine_optim_clusterWrap(cl = NULL, nodes, reps, objs, pkgs)

Arguments

cl

a cluster object or NULL

nodes

positive integer of length 1 specifying the number of threads to use on the current machine (localhost), or a character vector of hostnames on which to run worker copies.

reps

integer. number of replicates to run.

objs

character vector of names of objects to export

pkgs

character vector of packages to pre-load on the cluster nodes

Value

named list of length 2 containing: cl, a cluster object; out, a list of burn maps (aka burnMapList).